split JSPropsUpdater
